Cleanly separate burn-in, line clearance, and next-tetrimino handling
Add ability for caller of the game logic (in this case a placeholder reagent app) to
inject in a custom play-next-tetrimino-fn which will be super helpful for
accommodating special-case game dynamics like debouncing the side-effects associated
with introducing the next tetrimino.

E.g. in the case when last-moment adjustments to rotation and horizontal position are
desired, even though the current tetrimino in play is on the baseline or adjacent to
peaks within the game in progress.

The refactor has introduced some new bugs so more tests will be needed to ironing the
issues